Move the JSON function MVI code into opt_mvi_jsonfuncs.cc
opt_multi_valued_index.cc held two separate concerns: the index side
(how a value is encoded into the index, the access descriptor, the quick
select) and the predicate side (which JSON functions can be computed
from an MVI, which of their arguments holds the indexed expression, and
what to search the index for). Split the second one out.

Moved verbatim:

  get_mvi_index()
  collect_mvi_keys()
  Item_func_json_contains::get_mvi_access() and ::mvi_analyze()
  Item_func_json_overlaps::get_mvi_access() and ::mvi_analyze()
  add_mvi_access()

The only code change is that encode_mvi_key() is no longer static: it is
used both by Item_func_mvi_encode::val_str_ascii(), which stays, and by
collect_mvi_keys(), which moves. It is declared in
opt_multi_valued_index.h now. The other three moved helpers had no
callers outside the moved code and stay static.

Item_func_mvi_encode is not a JSON function and stays put: it is how the
values get into the index in the first place.
