+++ sql/opt_multi_valued_index.cc
@@ -87,8 +87,8 @@ static void store_sort_key_longlong(uchar *to, bool unsigned_flag,
   to[0]= (uchar) (value >> 56) ^ (unsigned_flag ? 0 : 128);
 }
 
-static bool encode_mvi_key(json_engine_t *je, const Type_handler *cast_th,
-                           CHARSET_INFO *cs, String *buf)
+bool encode_mvi_key(json_engine_t *je, const Type_handler *cast_th,
+                    CHARSET_INFO *cs, String *buf)
 {
   enum_field_types cast_ftype= cast_th->field_type();
   bool is_unsigned= cast_th->is_unsigned();
@@ -270,29 +270,6 @@ bool collect_mvi_vcols_for_join(JOIN *join, List<Mv_index> *indexes)
   return FALSE; // Ok
 }
 
-/*
-  Find Multi-Value Index created over array_indexed_expr.
-*/
-static Mv_index *get_mvi_index(List<Mv_index> *indexes,
-                               Item *array_indexed_expr)
-{
-  Mv_index *index;
-  List_iterator<Mv_index> it(*indexes);
-  Item_func_mvi_encode *mvitem;
-  while ((index= it++))
-  {
-    Field *vcol_field= index->vcol;
-    DBUG_ASSERT(vcol_field->vcol_info->expr->type() == Item::FUNC_ITEM);
-    DBUG_ASSERT(((Item_func *) vcol_field->vcol_info->expr)->functype() ==
-                Item_func::MVI_ENCODE_FUNC);
-    mvitem= (Item_func_mvi_encode *) vcol_field->vcol_info->expr;
-    if (mvitem->arguments()[0]->eq(array_indexed_expr, true))
-    {
-      return index;
-    }
-  }
-  return NULL;
-}
 
 /*
   Add one encoded element key to the access.
@@ -365,219 +342,6 @@ void Mvi_access::print_json(THD *thd, Json_writer_object *trace_object)
   Json_writer_array trace_ranges(thd, "ranges");
   while ((key= it++))
     trace_ranges.add(key->ptr(), key->length());
-}
-
-
-/*
-  @brief
-    Collect the element keys to search `index' for from a JSON literal.
-
-  @param cs           Collation of the indexed expression
-                      TODO why does that matter?
-
-  @param json         The JSON literal: an array, or a single scalar
-  @param conjunctive  true when the keys are ANDed (JSON_CONTAINS),
-                      false when they are ORed (JSON_OVERLAPS)
-  @param je           A json_engine_t to scan with
-
-  @detail
-    An element that cannot be encoded for this index (a type mismatch, say)
-    can only be skipped when the keys are ANDed. Dropping a key from an AND
-    makes the index scan less selective, so it still returns a superset of
-    the rows the predicate matches, and the predicate itself does the exact
-    filtering afterwards.
-
-    For an OR we cannot do that. A row can satisfy the predicate through the
-    very element we failed to encode, and MVI_ENCODE skips such elements too,
-    so that row has no key in the index for us to find it by. Dropping the
-    key would lose it. Give up on the access instead.
-
-  @return
-    The access descriptor, or NULL if the predicate cannot use this MVI.
-*/
-
-static Mvi_access *collect_mvi_keys(THD *thd, Mv_index *index,
-                                    CHARSET_INFO *cs, String *json,
-                                    bool conjunctive, json_engine_t *je)
-{
-  Mvi_access *access= NULL;
-  StringBuffer<256> buf;
-  const uchar *start= reinterpret_cast<const uchar *>(json->ptr());
-  const uchar *end= start + json->length();
-  Item_func_mvi_encode *mvitem=
-    (Item_func_mvi_encode *) index->vcol->vcol_info->expr;
-  const Type_handler *cast_th= mvitem->cast_type().type_handler();
-
-  buf.length(0);
-  buf.set_charset(&my_charset_latin1_bin);
-
-  if (json_scan_start(je, cs, start, end) || json_read_value(je))
-    return NULL;
-
-  if (je->value_type == JSON_VALUE_UNINITIALIZED ||
-      je->value_type == JSON_VALUE_OBJECT)
-    return NULL;
-
-  if (je->value_type != JSON_VALUE_ARRAY)
-  {
-    /* A scalar: JSON_CONTAINS(expr, '123') */
-    if (encode_mvi_key(je, cast_th, cs, &buf))
-      return NULL;
-    if (!(access= new (thd->mem_root) Mvi_access(index, conjunctive)) ||
-        access->add_key(thd->mem_root, &buf))
-      return NULL;
-    return access;
-  }
-  // JSON_VALUE_ARRAY
-
-  /* TODO: deduplicate? */
-  do {
-    buf.length(0);
-    switch (je->state)
-    {
-      /* TODO: nested array? */
-      case JST_ARRAY_START:
-        continue;
-      case JST_ARRAY_END:
-        break;
-      case JST_VALUE:
-      {
-        if (json_read_value(je))
-          return NULL;
-
-        if (encode_mvi_key(je, cast_th, cs, &buf))
-        {
-          /* See above: only an AND of the keys tolerates a missing one */
-          if (!conjunctive)
-            return NULL;
-          break;
-        }
-        if (!access &&
-            !(access= new (thd->mem_root) Mvi_access(index, conjunctive)))
-          return NULL;
-        if (access->add_key(thd->mem_root, &buf))
-          return NULL;
-        break;
-      }
-      default:
-        return NULL;
-    }
-  } while (json_scan_next(je) == 0);
-
-  return access;
-}
-
-
-/*
-  @brief
-    Check if we can use Multi-Value Index access to read rows for this
-    predicate, if yes create an access descriptor.
-
-  @detail
-    Check if this item is a
-
-      JSON_CONTAINS(array_indexed_expr, '[foo, bar, ... ]')
-
-    which is true when ALL of the elements have a match, so the keys are
-    ANDed.
-
-  @return
-    The access descriptor, or NULL if the predicate cannot use an MVI.
-*/
-
-Mvi_access *Item_func_json_contains::get_mvi_access(THD *thd,
-                                                    List<Mv_index> *indexes)
-{
-  Mv_index *index;
-  DBUG_ASSERT(fixed());
-
-  if (arg_count > 2 || !a2_constant)
-    return NULL;
-  /* Find the MVI that matches the first argument */
-  if (!(index= get_mvi_index(indexes, args[0])))
-    return NULL;
-
-  if (!a2_parsed)
-  {
-    val= args[1]->val_json(&tmp_val);
-    a2_parsed= true;
-  }
-  if (!val)
-    return NULL;
-
-  return collect_mvi_keys(thd, index, args[0]->collation.collation, val,
-                          true, &je);
-}
-
-
-/*
-  @brief
-    Check if we can use Multi-Value Index access to read rows for this
-    predicate, if yes create an access descriptor.
-  
-  @detail
-    We can use MVI index when the predicate has either of the forms:
-
-      JSON_OVERLAPS(array_indexed_expr, '[foo, bar, ... ]')
-      JSON_OVERLAPS('[foo, bar, ... ]', array_indexed_expr)
-
-    JSON_OVERLAPS is true when ANY of the elements has a match, so the keys
-    are ORed. 
-
-  @return
-    The access descriptor, or NULL if the predicate cannot use an MVI.
-*/
-
-Mvi_access *Item_func_json_overlaps::get_mvi_access(THD *thd,
-                                                    List<Mv_index> *indexes)
-{
-  Mv_index *index;
-  uint literal_arg;
-  String *json;
-  StringBuffer<256> tmp;
-  DBUG_ASSERT(fixed());
-
-  if ((index= get_mvi_index(indexes, args[0])))
-    literal_arg= 1;
-  else if ((index= get_mvi_index(indexes, args[1])))
-    literal_arg= 0;
-  else
-    return NULL;
-
-  if (!args[literal_arg]->const_item())
-    return NULL;
-  if (!(json= args[literal_arg]->val_json(&tmp)))
-    return NULL;
-
-  /*
-    encode_mvi_key() must see the collation of the indexed expression: that
-    is what decides how MVI_ENCODE built the keys that are in the index.
-  */
-  return collect_mvi_keys(thd, index,
-                          args[1 - literal_arg]->collation.collation, json,
-                          false, &je);
-}
-
-
-/* Add `access' to the context, if there is one. Returns true on error */
-
-static bool add_mvi_access(Mvi_context *ctx, Mvi_access *access)
-{
-  return access && ctx->accesses.push_back(access, ctx->thd->mem_root);
-}
-
-
-bool Item_func_json_contains::mvi_analyze(void *arg)
-{
-  Mvi_context *ctx= (Mvi_context *) arg;
-  return add_mvi_access(ctx, get_mvi_access(ctx->thd, &ctx->indexes));
-}
-
-
-bool Item_func_json_overlaps::mvi_analyze(void *arg)
-{
-  Mvi_context *ctx= (Mvi_context *) arg;
-  return add_mvi_access(ctx, get_mvi_access(ctx->thd, &ctx->indexes));
 }
